Fix merge

+++ vm/devices/virtio/virtio/src/device.rs
@@ -206,11 +206,11 @@ mod tests {
             DeviceTraits::default()
         }
 
-        fn read_registers_u32(&mut self, _offset: u16) -> u32 {
+        fn read_registers_u32(&mut self, _offset: u64) -> u32 {
             0
         }
 
-        fn write_registers_u32(&mut self, _offset: u16, _val: u32) {}
+        fn write_registers_u32(&mut self, _offset: u64, _val: u32) {}
 
         fn start_queue(
             &mut self,
